[WIP] support la32r interrupt & exception diff
When an interrupt or exception is encountered, nemu will change the csr state and execute the first instruction of exception handling in one cycle, and these operations Nutshell needs two difftest commits to complete, so use the difftestExceptionSkip signal to determine whether it is the difftest commit that "generated the int or excp" . If it is, skips the nemu exec and the register comparison. (This logic is cumbersome and probably needs to be refactored).